market-data02-consolidated/requirements.txt
numpy
matplotlib
yamal==8.0.5
msgpack
featuremine-extractor==7.0.3